+++ packages/blockchain-utils-linking/src/__tests__/near.test.ts
@@ -1,4 +1,5 @@
-import { NearAuthProvider } from '../near'
+import { jest } from '@jest/globals'
+import { NearAuthProvider } from '../near.js'
 import * as nearApiJs from 'near-api-js'
 
 const did = 'did:3:bafysdfwefwe'