+++ packages/blockchain-utils-linking/src/eosio.ts
@@ -1,7 +1,7 @@
-import { AuthProvider } from './auth-provider'
-import { AccountID } from 'caip'
-import { getConsentMessage, LinkProof } from './util'
-import { normalizeAccountId } from './ethereum'
+import { AuthProvider } from './auth-provider.js'
+import { AccountId } from 'caip'
+import { getConsentMessage, LinkProof } from './util.js'
+import { normalizeAccountId } from './ethereum.js'
 import * as sha256Stable from '@stablelib/sha256'
 import * as uint8arrays from 'uint8arrays'
 
@@ -12,9 +12,9 @@ export class EosioAuthProvider implements AuthProvider {
 
   constructor(private readonly provider: any, private readonly address: string) {}
 
-  async accountId(): Promise<AccountID> {
+  async accountId(): Promise<AccountId> {
     const chainId = toCAIPChainId(await this.provider.getChainId())
-    return new AccountID({
+    return new AccountId({
       address: this.address,
       chainId: `eosio:${chainId}`,
     })
@@ -55,7 +55,7 @@ function sanitize(str: string, size: number): string {
   return str.replace(/\s/g, ' ').replace(new RegExp(`(\\S{${size}})`, 'g'), '$1 ')
 }
 
-export function toPayload(message: string, accountID: AccountID): string {
+export function toPayload(message: string, accountID: AccountId): string {
   const { address, chainId } = accountID
   const payload = `${message} [For: ${address} on chain: ${chainId}]`
   return sanitize(payload, maxWordLength)
@@ -63,7 +63,7 @@ export function toPayload(message: string, accountID: AccountID): string {
 
 export async function toSignedPayload(
   message: string,
-  accountID: AccountID,
+  accountID: AccountId,
   provider: any
 ): Promise<string> {
   accountID = normalizeAccountId(accountID)