+++ packages/blockchain-utils-linking/src/polkadot.ts
@@ -1,6 +1,6 @@
-import { AccountID } from 'caip'
-import { AuthProvider } from './auth-provider'
-import { getConsentMessage, LinkProof } from './util'
+import { AccountId } from 'caip'
+import { AuthProvider } from './auth-provider.js'
+import { getConsentMessage, LinkProof } from './util.js'
 import * as uint8arrays from 'uint8arrays'
 
 const stringHex = (str: string): string =>
@@ -36,8 +36,8 @@ export class PolkadotAuthProvider implements AuthProvider {
     }
   }
 
-  async accountId(): Promise<AccountID> {
-    return new AccountID({ address: this.address, chainId: CHAIN_ID })
+  async accountId(): Promise<AccountId> {
+    return new AccountId({ address: this.address, chainId: CHAIN_ID })
   }
 
   withAddress(address: string): AuthProvider {