+++ packages/blockchain-utils-linking/src/tezos.ts
@@ -1,6 +1,6 @@
-import { AccountID } from 'caip'
-import { AuthProvider } from './auth-provider'
-import { getConsentMessage, LinkProof } from './util'
+import { AccountId } from 'caip'
+import { AuthProvider } from './auth-provider.js'
+import { getConsentMessage, LinkProof } from './util.js'
 import { hash } from '@stablelib/sha256'
 import * as uint8arrays from 'uint8arrays'
 import type { Signer } from '@taquito/taquito'
@@ -91,7 +91,7 @@ export class TezosAuthProvider implements AuthProvider {
     const signature = await sign(this.provider, message)
     const address = await getActiveAddress(this.provider)
     // generate account ID
-    const caipAccount = new AccountID({
+    const caipAccount = new AccountId({
       address,
       chainId: `${TEZOS_NAMESPACE}:${TEZOS_CHAIN_REF}`,
     })
@@ -109,9 +109,9 @@ export class TezosAuthProvider implements AuthProvider {
   /**
    * @inheritdoc
    */
-  async accountId(): Promise<AccountID> {
+  async accountId(): Promise<AccountId> {
     const address = await getActiveAddress(this.provider)
-    return new AccountID({
+    return new AccountId({
       address,
       chainId: `${TEZOS_NAMESPACE}:${TEZOS_CHAIN_REF}`,
     })