+++ packages/blockchain-utils-validation/src/blockchains/__tests__/cosmos.test.ts
@@ -1,17 +1,19 @@
-import { signTx, Tx, SignMeta, createWalletFromMnemonic, Wallet, StdTx } from '@tendermint/sig'
-import { validateLink } from '../cosmos'
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import tendermint from '@tendermint/sig'
+import { validateLink } from '../cosmos.js'
 import * as linking from '@ceramicnetwork/blockchain-utils-linking'
 
 const did = 'did:3:bafysdfwefwe'
 const mnemonic = 'test salon husband push melody usage fine ensure blade deal miss twin'
-const localProvider = createWalletFromMnemonic(mnemonic)
+const localProvider = tendermint.createWalletFromMnemonic(mnemonic)
 const chainRef = 'cosmoshub-3'
 
 class CosmosMockSigner {
-  constructor(readonly provider: Wallet) {}
+  constructor(readonly provider: tendermint.Wallet) {}
 
-  public async sign(msg: Tx, metadata: SignMeta): Promise<StdTx> {
-    return signTx(msg, metadata, this.provider)
+  public async sign(msg: tendermint.Tx, metadata: tendermint.SignMeta): Promise<tendermint.StdTx> {
+    return tendermint.signTx(msg, metadata, this.provider)
   }
 }