+++ packages/blockchain-utils-validation/src/blockchains/tezos.ts
@@ -1,9 +1,10 @@
-import { AccountID, ChainID } from 'caip'
+import { ChainId } from 'caip'
 import { RemoteSigner } from '@taquito/remote-signer'
-import { BlockchainHandler } from '../blockchain-handler'
+import { BlockchainHandler } from '../blockchain-handler.js'
 import { LinkProof, tezos } from '@ceramicnetwork/blockchain-utils-linking'
 import fetch from 'cross-fetch'
 import * as uint8arrays from 'uint8arrays'
+import { normalizeAccountId } from '@ceramicnetwork/common'
 
 export const ADDRESS_NOT_FOUND_ERROR = new Error(`Address not found on the Tezos blockchain`)
 export const PUBLIC_KEY_NOT_PUBLISHED_ERROR = new Error(
@@ -45,7 +46,7 @@ function encodeMessage(text: string): string {
  */
 function publicKeyFinder(address: string): () => Promise<string> {
   // request the public key from the Tezos blockchain
-  const request = fetch(`https://api.tzstats.com/explorer/account/${address}`).catch((e: any) => {
+  const request = fetch(`https://api.tzstats.com/explorer/account/${address}`).catch(() => {
     throw ADDRESS_NOT_FOUND_ERROR
   })
   return async (): Promise<string> => {
@@ -64,8 +65,8 @@ function publicKeyFinder(address: string): () => Promise<string> {
 // - at least one transaction on the blockchain
 // - their public key published to the blockchain
 export async function validateLink(proof: LinkProof): Promise<LinkProof | null> {
-  const account = AccountID.parse(proof.account)
-  const chainId = new ChainID(account.chainId)
+  const account = normalizeAccountId(proof.account)
+  const chainId = new ChainId(account.chainId)
 
   // only support Tezos mainnet for now
   if (chainId.reference !== TEZOS_CHAIN_REF) {
@@ -85,9 +86,7 @@ export async function validateLink(proof: LinkProof): Promise<LinkProof | null> 
   }
 }
 
-const Handler: BlockchainHandler = {
+export const handler: BlockchainHandler = {
   namespace: TEZOS_NAMESPACE,
   validateLink,
 }
-
-export default Handler