+++ packages/core/src/anchor/anchor-service-response.ts
@@ -1,9 +1,9 @@
-import CID from 'cids'
+import type { CID } from 'multiformats/cid'
 
 /**
  * Describes anchor service response
  */
-export default interface AnchorServiceResponse {
+export interface AnchorServiceResponse {
   readonly cid: CID
   readonly status: string
   readonly message: string