+++ packages/core/src/dispatcher.ts
@@ -1,4 +1,4 @@
-import CID from 'cids'
+import { CID } from 'multiformats/cid'
 import cloneDeep from 'lodash.clonedeep'
 import {
   DiagnosticsLogger,
@@ -7,21 +7,22 @@ import {
   StreamUtils,
   UnreachableCaseError,
 } from '@ceramicnetwork/common'
-import StreamID from '@ceramicnetwork/streamid'
-import { Repository } from './state-management/repository'
+import { StreamID } from '@ceramicnetwork/streamid'
+import { Repository } from './state-management/repository.js'
 import {
   MsgType,
   PubsubMessage,
   QueryMessage,
   ResponseMessage,
   UpdateMessage,
-} from './pubsub/pubsub-message'
-import { Pubsub } from './pubsub/pubsub'
+} from './pubsub/pubsub-message.js'
+import { Pubsub } from './pubsub/pubsub.js'
 import { Subscription } from 'rxjs'
-import { MessageBus } from './pubsub/message-bus'
-import { LRUMap } from 'lru_map'
-import { PubsubKeepalive } from './pubsub/pubsub-keepalive'
-import { PubsubRateLimit } from './pubsub/pubsub-ratelimit'
+import { MessageBus } from './pubsub/message-bus.js'
+import lru from 'lru_map'
+import { PubsubKeepalive } from './pubsub/pubsub-keepalive.js'
+import { PubsubRateLimit } from './pubsub/pubsub-ratelimit.js'
+import { TaskQueue } from './pubsub/task-queue.js'
 
 const IPFS_GET_RETRIES = 3
 const IPFS_GET_TIMEOUT = 30000 // 30 seconds per retry, 3 retries = 90 seconds total timeout
@@ -50,7 +51,7 @@ function messageTypeToString(type: MsgType): string {
  */
 export class Dispatcher {
   readonly messageBus: MessageBus
-  readonly dagNodeCache: LRUMap<string, any>
+  readonly dagNodeCache: lru.LRUMap<string, any>
   // Set of IDs for QUERY messages we have sent to the pub/sub topic but not yet heard a
   // corresponding RESPONSE message for. Maps the query ID to the primary StreamID we were querying for.
   constructor(
@@ -60,8 +61,16 @@ export class Dispatcher {
     private readonly _logger: DiagnosticsLogger,
     private readonly _pubsubLogger: ServiceLogger,
     maxQueriesPerSecond: number
+    readonly tasks: TaskQueue = new TaskQueue(),
   ) {
-    const pubsub = new Pubsub(_ipfs, topic, IPFS_RESUBSCRIBE_INTERVAL_DELAY, _pubsubLogger, _logger)
+    const pubsub = new Pubsub(
+      _ipfs,
+      topic,
+      IPFS_RESUBSCRIBE_INTERVAL_DELAY,
+      _pubsubLogger,
+      _logger,
+      tasks
+    )
     this.messageBus = new MessageBus(
       new PubsubRateLimit(
         new PubsubKeepalive(pubsub, MAX_PUBSUB_PUBLISH_INTERVAL),
@@ -70,7 +79,7 @@ export class Dispatcher {
       )
     )
     this.messageBus.subscribe(this.handleMessage.bind(this))
-    this.dagNodeCache = new LRUMap<string, any>(IPFS_CACHE_SIZE)
+    this.dagNodeCache = new lru.LRUMap<string, any>(IPFS_CACHE_SIZE)
   }
 
   /**
@@ -84,9 +93,14 @@ export class Dispatcher {
       if (StreamUtils.isSignedCommitContainer(data)) {
         const { jws, linkedBlock } = data
         // put the JWS into the ipfs dag
-        const cid = await this._ipfs.dag.put(jws, { format: 'dag-jose', hashAlg: 'sha2-256' })
+        const cid = await this._ipfs.dag.put(jws, { storeCodec: 'dag-jose', hashAlg: 'sha2-256' })
         // put the payload into the ipfs dag
-        await this._ipfs.block.put(linkedBlock, { cid: jws.link.toString() })
+        const linkCid = jws.link
+        const format = await this._ipfs.codecs.getCodec(linkCid.code).then((f) => f.name)
+        const mhtype = await this._ipfs.hashers
+          .getHasher(linkCid.multihash.code)
+          .then((mh) => mh.name)
+        await this._ipfs.block.put(linkedBlock, { format, mhtype, version: linkCid.version })
         await this._restrictCommitSize(jws.link.toString())
         await this._restrictCommitSize(cid)
         return cid
@@ -149,7 +163,7 @@ export class Dispatcher {
    * Helper function for loading a CID from IPFS
    */
   private async _getFromIpfs(cid: CID | string, path?: string): Promise<any> {
-    const asCid = typeof cid === 'string' ? new CID(cid) : cid
+    const asCid = typeof cid === 'string' ? CID.parse(cid) : cid
 
     // Lookup CID in cache before looking it up IPFS
     const cidAndPath = path ? asCid.toString() + path : asCid.toString()
@@ -172,7 +186,7 @@ export class Dispatcher {
           err.message == 'Request timed out'
         ) {
           console.warn(
-            `Timeout error while loading CID ${cid.toString()} from IPFS. ${retries} retries remain`
+            `Timeout error while loading CID ${asCid.toString()} from IPFS. ${retries} retries remain`
           )
           if (retries > 0) {
             continue
@@ -193,7 +207,8 @@ export class Dispatcher {
    * @private
    */
   async _restrictCommitSize(cid: CID | string): Promise<void> {
-    const stat = await this._ipfs.block.stat(cid, { timeout: IPFS_GET_TIMEOUT })
+    const asCid = typeof cid === 'string' ? CID.parse(cid) : cid
+    const stat = await this._ipfs.block.stat(asCid, { timeout: IPFS_GET_TIMEOUT })
     if (stat.size > IPFS_MAX_COMMIT_SIZE) {
       throw new Error(
         `${cid.toString()} commit size ${
@@ -309,6 +324,7 @@ export class Dispatcher {
    */
   async close(): Promise<void> {
     this.messageBus.unsubscribe()
+    await this.tasks.onIdle()
   }
 
   /**