+++ packages/core/src/state-management/named-task-queue.ts
@@ -1,4 +1,4 @@
-import { noop, TaskQueue } from '../pubsub/task-queue'
+import { noop, TaskQueue } from '../pubsub/task-queue.js'
 
 /**
  * Set of named PQueues.