+++ packages/http-client/src/__tests__/ceramic-http-client.test.ts
@@ -1,4 +1,4 @@
-import CID from 'cids'
+import { CID } from 'multiformats/cid'
 import {
   CeramicCommit,
   CommitType,
@@ -6,14 +6,14 @@ import {
   StreamUtils,
   TestUtils,
 } from '@ceramicnetwork/common'
-import { Document } from '../document'
-import CeramicClient from '../ceramic-http-client'
+import { Document } from '../document.js'
+import { CeramicClient } from '../ceramic-http-client.js'
 import { StreamID } from '@ceramicnetwork/streamid'
 import { TileDocument } from '@ceramicnetwork/stream-tile'
 
 const API_URL = 'https://example.com'
-const FAKE_CID_1 = new CID('bafybeig6xv5nwphfmvcnektpnojts33jqcuam7bmye2pb54adnrtccjlsu')
-const FAKE_CID_2 = new CID('bafybeig6xv5nwphfmvcnektpnojts44jqcuam7bmye2pb54adnrtccjlsu')
+const FAKE_CID_1 = CID.parse('bafybeig6xv5nwphfmvcnektpnojts33jqcuam7bmye2pb54adnrtccjlsu')
+const FAKE_CID_2 = CID.parse('bafybeig6xv5nwphfmvcnektpnojts44jqcuam7bmye2pb54adnrtccjlsu')
 const initial = {
   type: 0,
   log: [