+++ packages/pinning-powergate-backend/src/index.ts
@@ -1,4 +1,4 @@
-import type CID from 'cids'
+import type { CID } from 'multiformats/cid'
 import type { CidList, PinningBackend, PinningInfo } from '@ceramicnetwork/common'
 import { createPow, Pow } from '@textile/powergate-client'
 import * as sha256 from '@stablelib/sha256'