+++ packages/pkh-did-resolver/src/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import index from '../index'
+import * as index from '../index.js'
 import { Resolver } from 'did-resolver'
 
 const ethDid = 'did:pkh:eip155:1:0xab16a96d359ec26a11e2c2b3d8f8b8942d5bfcdb'
@@ -9,7 +9,6 @@ describe('PKH DID Resolver', () => {
   it('successfully resolves DIDs', async () => {
     const resolverRegistry = index.getResolver()
     const resolver = new Resolver(resolverRegistry)
-
 
     expect(await resolver.resolve(ethDid)).toMatchSnapshot()
     expect(await resolver.resolve(btcDid)).toMatchSnapshot()